Fix new password bug
This fixes a problem with password resets that occurred
on newly upgraded installations of CC.  By default, CC7 prompts
all users to upgrade their passwords.  If the user, for whatever
reason, presses forgot password, a pbk2df temp password is sent
to them; because CC7 is set to force_new_password, it runs md5
on the temp password and the user can't get in.  Now fixed.