BUGFIX: Fix disappearing documents
When a user cut and moved a folder, an unknown javascript bug would
occasionally cause the path value not to be registered. The server-side script
would then erroneously update all document paths for the current case,
making it appear to the user that all case documents had disappeared. This
fix throws an error if the path value is not registered client side
and prompts the user to reload page and try again.